Make sure OIDC/GraphQL client integration runs after client config is initialized, get rid of config merger bean
