+++ extensions/smallrye-graphql-client/deployment/src/main/java/io/quarkus/smallrye/graphql/client/deployment/GraphQLClientConfigInitializedBuildItem.java
@@ -1,0 +1,9 @@
+package io.quarkus.smallrye.graphql.client.deployment;
+
+import io.quarkus.builder.item.SimpleBuildItem;
+
+/**
+ * This marks that the SmallRye GraphQL Client configurations have been processed and initialized.
+ */
+public final class GraphQLClientConfigInitializedBuildItem extends SimpleBuildItem {
+}
